backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.4
pydantic-settings==2.2.1
paho-mqtt==1.6.1
httpx==0.27.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
